webapi/webapi-webrtc-w3c-tests/webrtc/RTCPeerConnection_createDTMFSender_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu,Yun <yunx.liu@intel.com>

-->

<meta charset="utf-8">
<title>WebRTC Test: RTCPeerConnection - createDTMFSender basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcpeerconnection-interface-extensions">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>

var t = async_test("Check if RTCPeerConnection.createDTMFSender exists and type of function", {timeout: 5000});

var pc;
var ds;

setup(function () {
  var configuration = {"iceServers": [{"url": "stun:stun.example.org"}]};
  if (typeof RTCPeerConnection != "undefined") {
    pc = new RTCPeerConnection(configuration);
  } else {
    pc = new webkitRTCPeerConnection(configuration);
  }
});

t.step(function () {
  if (typeof (navigator.getUserMedia) != "undefined") {
    navigator.getUserMedia({audio: true, video: true}, gotStream, error);
  } else {
    navigator.webkitGetUserMedia({audio: true, video: true}, gotStream, error);
  }
});

function gotStream(stream) {
  pc.addStream(stream);
  ds = pc.createDTMFSender(stream.getAudioTracks()[0]);

  t.step(function () {
    assert_true("createDTMFSender" in pc, "createDTMFSender attribute in RTCPeerConnection");
    assert_equals(typeof pc.createDTMFSender, "function", "createDTMFSender attribute of type");
  });
  t.done();
}

function error() {
  t.step(function () {
    assert_unreached("Stream generation failed");
  });
  t.done();
}

</script>
